tests: stop passing invalid ops to futex_op_wake
Kernel 4.14-rc1 (commit 30d6e0a4190d) started checking op parameter to
futex_op_wake and strace tests started failing due to that:
----8<--------8<--------8<--------8<--------8<----
FAIL: futex
===========

futex(0x7fabd78bcffc, 0x5, 0xfacefeed, 0xb, 0x7fabd78bcffc, 0xa0caffee) = -1: Invalid argument
futex.test: failed test: ../futex failed with code 1
----8<--------8<--------8<--------8<--------8<----

0xa0caffee decodes as:
0x80000000 & 0xa0caffee: oparg is shift
0x70000000 & 0xa0caffee: op is FUTEX_OP_OR
0x0f000000 & 0xa0caffee: cmp is FUTEX_OP_CMP_EQ
0x00fff000 & 0xa0caffee: oparg is sign-extended 0xcaf = -849
0x00000fff & 0xa0caffee: cmparg is sign-extended 0xfee = -18

That means the op tries to do this:
	(futex |= (1 << (-849))) == -18
which is completely bogus. The new check of op in the code is:
	if (encoded_op & (FUTEX_OP_OPARG_SHIFT << 28)) {
	    if (oparg < 0 || oparg > 31)
		    return -EINVAL;
	    oparg = 1 << oparg;
	}

which results obviously in the "Invalid argument" errno above.

The kernel will soften the failure to print only a (ratelimited)
message, but the tests should be fixed so that we can enforce the error
in the future again.

Signed-off-by: Jiri Slaby <jslaby@suse.cz>