strace: move pathtrace_select calls to a later initialisation stage
Since path name resolution is specific to the tracing backend
(especially when tracees are remote), let's first collect the list
of paths to trace in a temporary array and add it later during the
initialisation.

* strace.c (init): Add pathtrace_paths temporary array, with its
element count stored in pathtrace_count and size in pathtrace_size.
(init) <case 'P'>: Store argument in pathtrace_paths, grow it if needed.
(init): Iterate over pathtrace_paths and call pathtrace_select for each
element.  Free the array.
* tests/options-syntax.test: Check that path name resolution happens
but not too early.

Co-authored-by: Dmitry V. Levin <ldv@altlinux.org>