count.c: use %9.u conversion specification
Instead of printing to a temporary string.  This also fixes a subtle bug
when the output for a billion errors and more is cut and shown
incorrectly.

* count.c (struct call_counts): Change type of calls and errors fields
to unsigned int since they are used and printed as unsigned anyway.
(call_summary_pers): Remove error_str.  Print cc->errors
and error_cum using %9.u conversion specification.