linux/sparc{,64}: replace empty syscall entries with comments
* linux/sparc/syscallent.h ([52]): Remove, add a comment that it is
memory_ordering on sparc64.
([164]): Remove, add a comment that it is utrap_install on sparc64.
* linux/sparc64/syscallent.h ([31]): Remove, add a comment that it is
lchown32 on sparc32.
([32]): Remove, add a comment that it is fchown32 on sparc32.
([35]): Remove, add a comment that it is chown32 on sparc32.
([44]): Remove, add a comment that it is getuid32 on sparc32.
([53]): Remove, add a comment that it is getgid32 on sparc32.
([56]): Remove, add a comment that it is mmap2 on sparc32.
([69]): Remove, add a comment that it is geteuid32 on sparc32.
([70]): Remove, add a comment that it is getegid32 on sparc32.
([72]): Remove, add a comment that it is setreuid32 on sparc32.
([77]): Remove, add a comment that it is truncate64 on sparc32.
([82]): Remove, add a comment that it is setgroups32 on sparc32.
([84]): Remove, add a comment that it is ftruncate64 on sparc32.
([87]): Remove, add a comment that it is setuid32 on sparc32.
([89]): Remove, add a comment that it is setgid32 on sparc32.
([91]): Remove, add a comment that it is setfsuid32 on sparc32.
([94]): Remove, add a comment that it is setfsgid32 on sparc32.
([112]): Remove, add a comment that it is setregid32 on sparc32.
([115]): Remove, add a comment that it is getgroups32 on sparc32.
([155]): Remove, add a comment that it is fcntl64 on sparc32.
([231]): Remove, add a comment that it is time on sparc32.