strace.c: pre-allocate tcb on PTRACE_EVENT_{FORK,VFORK,CLONE}
In an attempt to resolve a race between the first stop of a new child
and the death of the last process.

* defs.h (TCB_PREALLOCATED): New constant.
* strace.c (pid2tcb): Add skip_preallocated parameter.
(maybe_allocate_tcb): Add tcp parameter, use it instead of allocating a
new tcp if it is not NULL.
(maybe_switch_tcbs): Pass true as a second argument to pid2tcb.
(next_event): Pass tcp to maybe_switch_tcbs if it is pre-allocated,
pre-allocate a tcb on PTRACE_EVENT_{FORK,VFORK,CLONE} if followfork is
enabled and we successfully retrieved the new pid.