ioctl: do not print comments twice in -Xverbose
Figure out whether the ioctl code is decoded inside a comment and adjust
printflags/printxval calls accordingly.

* ioctl.c (ioctl_print_code, evdev_decode_number): Add abbrev variable,
set it to true if xlat style is not XLAT_STYLE_VERBOSE, do not provide
dflt and set xlat style to XLAT_STYLE_ABBREV in printflags/printxval
calls (that are now changed to printflags_ex/printxval_ex to accomodate
the change).