filter_seccomp: fix jump offset overflow in binary match generator
binary_match_filter_generator() is missing a check for jump offset
overflows which might result in incorrect behavior if the binary match
strategy is selected and overflows.

I have only been able to reproduce the bug on mips after forcing strace to
use the binary match generator.  Due to the large number of syscalls on
mips, the binary match algorithm is suboptimal and the linear one is
selected.  This bug could however be triggered inadvertently if tracing a
very large set of syscalls not grouped together; in that case, the linear
strategy might have a jump offset overflow itself and strace would
fallback to the binary match one.

* filter_seccomp (binary_match_filter_generator): Check for jump offset
overflows.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>