tests: skip stack tracing tests when /proc/self/maps is invalid
Apparently, some container orchestration systems are capable of messing
up with /proc/self/maps.  For example, one can face with the following
garbage in Travis CI:

	$ cat /proc/self/maps
	ee72dcc0000-ee72dcd0000 r-xp 00000000 00:43 16 /var/snap/lxd/common/lxd/storage-pools/instances/containers/travis-job-strace-strace-634261782/rootfs/bin/cat
	ee72dcd0000-ee72dce0000 r--p 00000000 00:43 16 /var/snap/lxd/common/lxd/storage-pools/instances/containers/travis-job-strace-strace-634261782/rootfs/bin/cat
	ee72dce0000-ee72dcf0000 rw-p 00010000 00:43 16 /var/snap/lxd/common/lxd/storage-pools/instances/containers/travis-job-strace-strace-634261782/rootfs/bin/cat
	...

Our stack tracing cannot work properly when /proc/self/maps is damaged
this way.

* tests/strace-k.test: Check that the path to executable recorded
in /proc/self/maps is valid and skip the test if it's not the case.