strace: capitalise placeholders in option declarations in usage
It is already capitalised in descriptions (rather inconsistently,
however), the capitalisation is already used for "PROG" and "ARGS",
and it allows consistently distinguish placeholders from literals.
The manual page does not need the same treatment, it seems,
as the italic script is used there for that purpose.

Capitalise the "usage:" part, as it is often capitalised in the usage
messages of some other programs.

* strace.c (usage): Capitalise "expr", "column", "interruptible",
"path", "file", "var", "val", "strsize", "pid", "format", "overhead",
"sortby" placeholders where appropriate. Capitalise the first letter
in "usage:".