tests/syntax: fix incorrect match_diff/macth_grep usage
match_diff and match_grep take the error message as the third argument
and pass it on to fail_, while the calling code expects that they simply
return non-zero in case of error.

* tests/syntax.sh (check_exit_status_and_stderr,
check_exit_status_and_stderr_using_grep): Specify the error message as
the third argument to match_diff/match grep instead of trying to handle
the erroneous status code.