strace: print long options in the option error message
If a long option alias is used with incorrect argument, it's better show it
in the error message instead of the short option equivalent.

* strace.c (error_opt_arg): Add lopt argument, print its name
if it is non-NULL.
(init): Add lopt_idx variable, pass it to getopt_long, set loop-local
lopt variable to an option pointer based on lopt_idx; pass lopt
in error_opt_arg calls.
* tests/options-syntax.test: Update expected output.