tests/xetitimer: Use syscalls instead of libc functions
The new glibc implementation for y2038 safe arches (32-bit arch with
64-bit time_t) will set the itimerval struct inside libc instead of
inside the kernel. This is because the kernel expects a 32-bit version
of timeval while libc exposes a 64-bit version. As libc sets this, when
we pass a faulting pointer we will get a segfault instead of the kernel
reporint EFAULT back.

To avoid a seg fault let's use the kernel's equivalent of struct
__kernel_old_itimerval in the tests and call the syscall directly
instead of using libc. This way we can trigger the kernels EFAULT
response correctly.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>

* tests/xetitimer.c: Include "kernel_timeval.h"
(kernel_old_itimerval_t): New type.
(main): Use it instead of struct itimerval, replace getitimer with
syscall(__NR_getitimer), replace setitimer with syscall(__NR_setitimer).