tests: disable umovestr_cached.test on mips o32 unless PTRACE_GET_SYSCALL_INFO works
If PTRACE_GET_SYSCALL_INFO does not work, then the test has to be
skipped on mips o32 because only the first four syscall arguments are
obtained from registers on that architecture, remaining syscall
arguments have to be fetched from the tracee memory.

* tests/umovestr_cached.test [$STRACE_ARCH == mips && $MIPS_ABI == o32]:
Skip the test if PTRACE_GET_SYSCALL_INFO does not work.