tests: introduce create_and_enter_subdir and leave_and_remove_subdir
Introduce the pair of functions that is going to be used to make sure
the current workdir of the tracee is different from the current workdir
of the tracer.

Use the new interface in a few tests where the difference is going to be
relevant with --secontext option.

* tests/subdir.c: New file.
* tests/Makefile.am (libtests_a_SOURCES): Add subdir.c.
* tests/tests.h (create_and_enter_subdir, leave_and_remove_subdir):
New function prototypes.
* tests/access.c (main): Use create_and_enter_subdir and
leave_and_remove_subdir.
* tests/chmod.c (main): Likewise.
* tests/execve.c (main): Likewise.
* tests/fchmod.c (main): Likewise.
* tests/fchmodat.c (main): Likewise.
* tests/fchownat.c (main): Likewise.
* tests/linkat.c (main): Likewise.
* tests/open.c (main): Likewise.
* tests/openat.c (main): Likewise.