netlink, nlattr: print structures followed by payloads as arrays
When struct nlmsgerr is followed by payload, print this sequence of data
as an array rather than a compound structure with a few nameless fields.

Likewise, when struct nlattr is followed by payload, print this sequence
of data as an array rather than a compound structure with nameless
fields.

* src/netlink.c (decode_nlmsgerr): Print the sequence of data
as an array when len > payload.
* src/nlattr.c (decode_nlattr_with_data): Use
tprint_array_begin and tprint_array_end instead of
tprint_struct_begin and tprint_struct_end
when nla_len > NLA_HDRLEN.
* NEWS: Mention this change.
* tests/nlattr.c (test_nlattr): Update expected output.
* tests/nlattr_ifinfomsg.c (print_prop_list_msg): Likewise.
* tests/nlattr_ifla.h (print_ifinfomsg): Likewise.
* tests/nlattr_ifla_af_spec.c (print_##family_##_msg): Likewise.
* tests/nlattr_ifla_linkinfo.c (TEST_UNKNOWN_TUNNELS, TEST_LINKINFO_,
TEST_NESTED_LINKINFO): Likewise.
* tests/nlattr_mdba_mdb_entry.c (print_br_port_msg, main): Likewise.
* tests/nlattr_mdba_router_port.c (print_br_port_msg, main): Likewise.
* tests/nlattr_nlmsgerr.c (print_nlmsgerr, main): Likewise.
* tests/nlattr_tc_stats.c (print_tcmsg): Likewise.
* tests/nlattr_tca_stab.c (print_tcmsg): Likewise.
* tests/test_nlattr.h (print_nlattr, print_sockfd): Likewise.