rtnl_link: pass 0 as opaque_data in decode_ifla_af_spec's decode_nlattr call
It has to be 0 to properly trigger decode_nlattr's size=0 special case
logic (since opaque_data argument is generated by decode_nlattr in that case),
but it was sloppily just passed through the caller's opaque_data
instead, which coincidentally used to be 0 up to v5.6~125 "rtnl_link:
use the family from ifinfomsg".

* src/rtnl_link.c (decode_ifla_af_spec): Pass 0 as the opaque_data argument
to the decode_nlattr() call.

Reported-by: Dmitry V. Levin <ldv@strace.io>
Fixes: v4.23~41 "rtnl_link: decode IFLA_AF_SPEC"
Fixes: v5.6~125 "rtnl_link: use the family from ifinfomsg"