close_range: fd and max_fd are unsigned
The distinction is important, because close_range(2) says that having
max_fd < fd is an error, but if you give the kernel 1U + INT_MAX as
max_fd, it'll interpret it as unsigned.

Since we're working with ranges of file descriptors here, the
particular properties of only the start and end file descriptors
aren't really relevant, so it doesn't matter that we're losing fancy
file descriptor printing in making this change.