tests: additional diagnostics in sockopt-timestamp
As of now, it prints only the actual cmsg_level/cmsg_type (that are
also likely to be print by strace itself) and not the expected ones;
since it is the purposely incorrect output that is not supposed to
match, add the expected cmsg_level/cmsg_type as well to aid debugging.

* tests/sockopt-timestamp.c (test_sockopt): Print expected cmsg_level
and cmsg_type in an "expected ..." comment.