tests: handle "-etrace=*" in addition to "-e trace=*" and "--trace=*"
Adjust the logic that recognises custom syscall tracing sets to include
"-etrace=..." syntax as well as "-e trace=..." and "--trace=..." in
order to allow uniform use of -e options within the sets of test
descriptions.

* tests/init.sh (run_strace_match_diff, run_strace_match_grep): Handle
"-etrace=*" in addition to "-e trace=*" and "--trace=*" as a disablement
of "-e trace=$NAME" heuristics.