trie: cache fill_value value instead of calculating it each time
trie_create_data_block may be a rather hot function if a trie is actively
used;  it seems wasteful to re-calculate fill_value each time a new data
block is allocated, considering the fact that it, as well as empty_value,
remains constant after the trie creation.  Let's prematurely optimise it
by pre-calculating the value in trie_create.

* src/trie.h (struct trie): Add fill_value field.
* src/trie.c (trie_create_data_block): Move fill_value calculation...
(trie_create): ...here, assign it to the fill_value field.