tests: reduce thoroughness of qual_fault's when= check
The addition of the third inner loop to qual_fault test has pushed
the test's execution time to the hight tens of seconds even on machines
with fast fork/exec.  Let's reduce its thoroughness a bit by reducing
the number of variants of first/last/step values being checked from six
down to four;  that should decrease the number of check_fault_injection
calls from 3040 down to 1200, a circa two-and-a-half-fold reduction.

* tests/qual_fault.test: Remove two elements from each of the iteration
lists for F, L, and S variables.