Extend -Y option to decode PIDs in arguments and return values
This change extends -Y/--decode-pids=comm option.  In addition to the
information printed before this change, command names would be printed
also for PIDs appeared in syscall arguments and return values.

For example:

  getppid()                  = 3781395<strace>
  getpid()                   = 3781398<a.out>
  pidfd_open(1<systemd>, 0)  = 3<pid:1<systemd>>

NOTE: The code printing the pid in printclockname() is not changed in
this commit.  CLOCKID_TO_FD macro used in printclockname() was removed
from the kernel, so printclockname() may change drastically in the
near future.

* src/defs.h (maybe_printpid_comm): New function declaration.
* src/strace.c (maybe_printpid_comm): New function printing
command name for the given PID.
* src/pidns.c (printpid_translation): Call `maybe_printpid_comm'.
* tests/strace--decode-pids-comm.c (do_default_action): Update expected
output.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>
Co-authored-by: Dmitry V. Levin <ldv@strace.io>