util: handle QUOTE_0_TERMINATED|QUOTE_EXPECT_TRAILING_0 in printstr_ex
There is a number of places in Linux kernel where the last character
of a string-typed structure field is not fetched and/or overwritten
with '\0'.  So far, it was difficult to signify this in strace output
performed via printstr_ex.  This change adds a logic that decreases
the size passed to print_quoted_string in order to signify the fact
that despite the presence of an additional non-NUL character in the field,
this non-NUL character will be ignored by the kernel.

* src/util.c (printstr_ex): Decrease size before print_quoted_string
call if both QUOTE_0_TERMINATED and QUOTE_EXPECT_TRAILING_0 set.