pidns: reorder struct translate_id_params fields
Moving output fields after input ones reduces the structure size
by whopping 25 percents (from 32 down to 24 bytes) on 64-bit systems,
and also makes the order of the fields a bit more natural.

* src/pidns.c (struct translate_id_params) <result_id, pd>: Move
to the end of the structure.
(translate_pid) <tip>: Reorder field initialisation in order
with the field order.