doc/strace.1: elaborate on support of multiple PIDs as a -p argument
While it was mentioned that -p "`pidof PROG`" syntax is supported,
it was not explicitly stated what characters can be used as separators
(like newline, which is used by pgrep, as noted in process_opt_p_list).

* doc/strace.1.in (.SS Startup): Explicitly mention that multiple PIDs
can be provided as an argument to a single -p option;  list all
supported separator characters;  mention possible -p "$(pgrep PROG)"
syntax.