Reduce sysctl decoder to the bare minimum
Given that sysctl syscall was removed by Linux kernel commit
v5.5-rc1~102^2, do not spend limited resources on the incomplete decoder
of sysctl syscall and reduce it to the bare minimum.  At the same time,
add mpers support and a test for the new decoder to make sure it works
properly.

* src/sysctl.c: Rewrite.
* src/xlat/sysctl_*.in: Remove.
* tests/sysctl.c: New file.
* tests/gen_tests.in (sysctl): New test.
* tests/pure_executables.list: Add sysctl.
* tests/.gitignore: Likewise.