hppa: Fix display of mask for rt_sigreturn syscall
Fix FUNC_GET_RT_SIGFRAME_ADDR to return correct address to the
sigframe so that the mask can be shown for the rt_sigreturn syscall.

This patch cleans up and simplifies the address calculation when
running on Linux kernels < 5.18, where the signal trampoline code
was stored in the executable stack.

With kernel 5.18 we added a VDSO with the trampoline code, could drop
the neccessarity of an executable stack and provided the offset to the
ucontext of the rt_sigframe struct for gdb and to simplify the
calculation sigframe address.

With this patch the rt_sigreturn.gen.test passes for 5.16 and 6.1
kernels on hppa.

Signed-off-by: Helge Deller <deller@gmx.de>