hppa: Fix pc.test testcase on hppa architecture
On hppa the lowest 2 bits of an address encodes the (parisc-specific)
address space in which a program runs. For userspace processes the space
number is 3, which is why an instruction at e.g. 0x100 shows up in the
address registers as 0x103.
This behaviour is different than on other architectures, which is
why the pc.test testcase fails on hppa.

Fix the testcase by simply dropping the last nibble of the address
when comparing the output, e.g. search for "0x10?" instead of "0x100".

Signed-off-by: Helge Deller <deller@gmx.de>