Add a configure flag for installing tests
--enable-install-tests=yes installs binaries and scripts
from the test suite to $libexec/strace.

This flag does not install everything necessary to run the test suite
(no harness), but at least installs the binaries and the scripts.

This can be useful when debugging various targets and how strace
behaves on them.

This can also be extended later with a standalone test harness
that would allow for an easier testing.

* configure.ac: Add --enable-install-tests.
* tests/Makefile.am [ENABLE_INSTALL_TESTS] (testslibexecdir,
testslibexec_PROGRAMS, testslibexec_DATA, testslibexec_SCRIPTS): New variables.

Signed-off-by: Elvira Khabirova <lineprinter0@gmail.com>