mips: fix raw_syscall_0
Some of the tests use a custom function (raw_syscall_0)
for syscall invocation.

The implementation for mips was flawed: it unconditionally
clobbered "hi" and "lo", which is not valid for MIPS R6,
because these targets lack those registers w/o DSP extension.

GCC 10+ refuses to compile such constructs: [1], [2].

Fix the syscall invocation implementation for mips.

[1] https://gcc.gnu.org/git/?p=gcc.git&a=commit;h=0a59215131c02dee4c8829f93d1ee678647614da
[2] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=91233

* src/linux/mips/raw_syscall.h (SYSCALL_CLOBBERLIST): Define.
(raw_syscall_0): Use it.

Signed-off-by: Elvira Khabirova <lineprinter0@gmail.com>